Registered parameter types go in their own files
